% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochroner.R
\name{get_query}
\alias{get_query}
\title{envoyer la requete}
\usage{
get_query(
  entry,
  q,
  costType,
  profile,
  timeUnit,
  distanceUnit,
  direction,
  resource,
  constraints
)
}
\arguments{
\item{entry}{un point dont les coordonnées sont exprimées dans le crs 4326
sous la forme "x,y" (transformé avec transform_sf)}

\item{q}{Une quantité de temps ou de distance (10 par défaut).}

\item{costType}{Précise le type de calcul qui est fait : "time" pour les isochrone
ou "distance" pour les isodistance (isochrones par défaut)}

\item{profile}{Type du coût utilisé pour le calcul : "pedestrian" pour les
piétons et "car" pour les voitures (Piéton par défaut).}

\item{timeUnit}{Permet de préciser l’unité dans laquelle les durées sont
exprimées dans la réponse : "hour", "minute" ou "second" (minutes par défaut)}

\item{distanceUnit}{Permet de préciser l’unité dans laquelle les distances
sont exprimées dans la réponse : "meter" ou "kilometer" (meter par défaut)}

\item{direction}{Sens du parcours. Soit on définit un point de départ
("departure") et on obtient les points d'arrivée potentiels.
Soit on définit un point d'arrivée ("arrival") et on obtient les points de
départ potentiels (departure par défaut)}

\item{resource}{Ce paramètre permet de préciser quelle ressource sera utilisée
pour le calcul. Actuellement, les ressources de type " ISO " et " PGR "
sont disponibles pour l’isochrone(par défaut "bdtopo-iso"). Voir section \code{RESOURCE}}

\item{constraints}{Permet d’exprimer des contraintes sur les caractéristiques
du réseau routier pour le calcul d’isochrones/isodistances. Voir section \code{CONSTRAINTS}.}
}
\value{
les résultats d'une requête
}
\description{
envoyer la requete
}
