% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochroner.R
\name{transform_sf}
\alias{transform_sf}
\title{transforme lse objets sf pour pouvoir les envoyer dans la requête}
\usage{
transform_sf(sf_object)
}
\arguments{
\item{sf_object}{une objet sf avec une géométrie de type POINT}
}
\value{
une liste de points dans le bon format pour la requête
}
\description{
transforme lse objets sf pour pouvoir les envoyer dans la requête
}
